Enhanced debugging support (object creation time and hooks, changed DebugName to Tag), and minor fixes to avoid warnings.

+++ trunk/source/direct3d10/Mesh.cpp
@@ -290,7 +290,7 @@ namespace Direct3D10
 		else 
 		{
 			faceRemap = gcnew array<int>( FaceCount );
-			int size = nativeFaceRemap.size();
+			int size = (int) nativeFaceRemap.size();
 			for( int faceIndex = 0; faceIndex < size; ++faceIndex )
 				faceRemap[faceIndex] = static_cast<int>( nativeFaceRemap[faceIndex] );
